Rename player to human
